Add 'Fix tool texture alignment' feature
This adds a feature that fixes the texture alignment for all tool
textures. This is done by calculating the texture axes based on the
normal vector of the brush face. It works the same way as using the face
alignment checkbox in the 'Face Edit Sheet' dialog in Hammer.

Background:

Texture axis (UV) information for tool textures is not preserved in the
BSP file, and the information read from the BSP file often results in
tool textures that have a perpendicular UV with respect to the
orientation of the brush face. These cases were usually caught and fixed
by the fixPerpendicular() method. However, in cases where the UV was not
perpendicular, the tool textures would sometimes appear excessively
stretched in the decompiled map.