Enable RenderInFastReflection for every game using DStaticPropV10CSGO
Cs:go engine branch no longer uses the 'screenSpaceFade' flag of
static_props for ScreenSpaceFade, but rather for
'RenderInFastReflection'. This change was already implemented, but only
used if the actual game was csgo. This commit fixes that the change is
also used for other games that use csgo's DStaticPropV10CSGO.