Simplify gui for areaportal/occluder mapping
This commit slightly simplifies the gui for forcing the mapping mode of
areaportals and occluders. Previously you could choose between Ordered
and Manual. However, because 'Ordered' is automatically selected in the
only scenario where it gives good results, there is no reason to force
it.
I removed and option and now you can only force
the manual mapping method.