Add 'smart' extract option
This adds the option to skip irrelevant files when extracting embedded
files. 'Irrelevant' files in this case are patched materials or
cubemap data generated by vBsp and only need by the engine to run the
map. These files are in almost any case useless to the normal user and
just create unnecessary clutter.