Add visgroup sorting by name
This commit implements a custom sorting algorithm for visgroups.
Visgroup names now get sorted by a 'AlphanumComparator', which tries to
implement a more 'human logic' sorting compared to the standard
string ascii sorting algorithm.
Internally this is done by splitting the string into multiple chunks,
where a chunk can either be a digit or a non digit.

Even though hammer really likes to remove empty visgroups, this commit
also implements that non used visgroups are not written to the final
vmf.

Additionally this commit changes, that instance visgroups are all
placed as childs of a parent 'Instances' visgroup. This makes it easier
to hide all instances at once.