Refactor the tooltexture fixing process
The new process delegates the tooltexture fixing to a new class
ToolTextureMatcher. This logic could even be abstracted away into an
interface in the future, enabling game specific tooltexture matchers.
This could be helpful for games which need some very specific special
handling for this process.

In best case scenarios the ToolTextureMatcher can now respect
surfaceproperties of materials.