Change version to 1.4.2-SNAPSHOT
Previously, the version number was always updated to specific <version>-SNAPSHOT when <version> was released. Normally <version>-SNAPSHOT is used before <version> is released so that on release, the version can be changed to just <version>. After that, you would immediately increase it to the next snapshot version.

Practically this doesn't make a difference for this project, as no maven artifacts are released, but it should be a little more intuitive to read which version is currently being "developed".