Don't overwrite user choice of fixEntityRot
Previously fixEntityRot only took effect if the bsp file version >= 21. This was done because fixEntityRot is (probably) only needed when the bsp was compiled with instances.
However, instances were already introduced in earlier bsp versions and are additionally able to be enabled in even earlier versions.
This commit simply ignores the bsp file version and just leaves the choice of fixing entity rotations to the user.
Motivated by #184